Corrected mocks for tests
Closes #1190
