+++ packages/core/src/service/__tests__/keyring-mock-data.ts
@@ -3,10 +3,10 @@
  * Generated from keyring-boot-mnemonics.test.ts
  */
 
+import { CURRENT_ID_KEY, KEYRING_STATE_V3_KEY } from '@onflow/flow-wallet-data-model';
 import encryptor from 'browser-passworder';
 
 import { SIGN_ALGO_NUM_ECDSA_P256 } from '@onflow/flow-wallet-shared/constant';
-import { CURRENT_ID_KEY, KEYRING_STATE_V3_KEY } from '@onflow/flow-wallet-shared/types';
 
 /**
  * Constants used for keyring testing
