+++ packages/core/src/service/userWallet.ts
@@ -69,6 +69,8 @@ import {
   tupleToPrivateKey,
 } from '@onflow/flow-wallet-shared/utils';
 
+import { getAccountsByPublicKeyTuple } from '@/utils/modules/findAddressWithPubKey';
+
 import { authenticationService } from '.';
 import keyringService from './keyring';
 import { mixpanelTrack } from './mixpanel';
@@ -78,7 +80,7 @@ import remoteConfigService from './remoteConfig';
 import transactionActivityService from './transaction-activity';
 import { defaultAccountKey, pubKeyAccountToAccountKey } from '../utils/account-key';
 import { fclConfig, fclConfirmNetwork } from '../utils/fclConfig';
-import { fetchAccountsByPublicKey, getAccountsByPublicKeyTuple } from '../utils/key-indexer';
+import { fetchAccountsByPublicKey } from '../utils/key-indexer';
 import {
   pk2PubKeyTuple,
   seed2PublicPrivateKey,
